siakad/requirements.txt
Flask>=3.0.0,<3.1
Flask-Login>=0.6.3,<0.7
Flask-SQLAlchemy>=3.1.1,<4.0
PyMySQL>=1.1.0,<2.0
python-dotenv>=1.0.1,<2.0
